- refactor friends into a different folder
