+++ align/align/urls.py
@@ -20,6 +20,7 @@ from rest_framework import routers
 from users import views as user_views
 from posts import views as posts_views
 from comments import views as comments_views
+from friends import views as friends_views
 
 router = routers.DefaultRouter()
 router.register(r'users', user_views.UserViewSet)
@@ -28,8 +29,8 @@ router.register(r'groups', user_views.GroupViewSet)
 router.register(r'posts', posts_views.PostsViewSet)
 router.register(r'comments', comments_views.CommentsViewSet)
 
-router.register(r'friendrequest', user_views.FriendRequestViewSet)
-router.register(r'author', user_views.IsFriendViewSet, basename="author")
+router.register(r'friendrequest', friends_views.FriendRequestViewSet)
+router.register(r'author', friends_views.IsFriendViewSet, basename="author")
 
 
 urlpatterns = [
