+++ align/friends/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from users.models import ExtendAuthorModel
+
+admin.site.register(ExtendAuthorModel)
+# Register your models here.
