+++ align/users/models.py
@@ -15,8 +15,3 @@ class User(AbstractUser):
     lastName = models.CharField(max_length=20, blank=True)
     displayName = models.CharField(max_length=40, blank=True, default="{} {}".format(firstName, lastName))
     github = models.URLField(blank=True)
-
-class ExtendAuthorModel(models.Model):
-    # change to reference host instead of id
-    authorID = models.ForeignKey(User, on_delete=models.CASCADE, related_name='authorID')
-    friendID = models.ForeignKey(User, on_delete=models.CASCADE, related_name='friendID')
